rename xlucene test file to be consistent
